Fix type-cast of pointer in buffer alignment check
In valign4(), which is intended to check if pointers are aligned, a debug
feature (activated with -DVERIFY_ALIGNMENT), pointers were naively cast to
unsigned integers.

This commit changes the cast of pointers to unsigned long, which should
work without compiler warnings on both 32-bit and 64-bit archs.

+++ src/openvpn/buffer.c
@@ -925,7 +925,7 @@ valign4 (const struct buffer *buf, const char *file, const int line)
   if (buf && buf->len)
     {
       int msglevel = D_ALIGN_DEBUG;
-      const unsigned int u = (unsigned int) BPTR (buf);
+      const unsigned long u = (unsigned long) BPTR (buf);
 
       if (u & (PAYLOAD_ALIGN-1))
 	msglevel = D_ALIGN_ERRORS;
